+++ open_web_calendar/conversion_base.py
@@ -16,12 +16,12 @@ from urllib.parse import urljoin
 import requests
 from lxml import etree
 
-from open_web_calendar.calendars import (
+from open_web_calendar.calendars.base import Calendars
+from open_web_calendar.calendars.caldav import (
     CalDAVCalendars,
-    Calendars,
-    ICSCalendars,
-    InvalidCalendars,
 )
+from open_web_calendar.calendars.errors import InvalidCalendars
+from open_web_calendar.calendars.ics import ICSCalendars
 from open_web_calendar.encryption import EmptyFernetStore, FernetStore
 
 
@@ -110,7 +110,7 @@ class ConversionStrategy:
             if not result:
                 # We did not find any link to any calendar from this source.
                 # We could have a CalDAV calendar.
-                result = CalDAVCalendars.from_url()
+                result = CalDAVCalendars.from_url(url)
                 if not result:
                     raise
             return result
